refactor(inspect): extract _ModuleSentinel to shared leaf module (#345)
Move _ModuleSentinel out of mcp/operations/inspect.py into a new leaf
module src/pyeye/_module_sentinel.py (public name: ModuleSentinel).
inspect.py imports it back as _ModuleSentinel via an alias so the single
call site (and all docstring references) are unchanged.

The leaf depends only on ast/pathlib at runtime; JediAnalyzer is
imported under TYPE_CHECKING only. This mirrors the _ast_targets.py
precedent and breaks the inspect→edges cycle: edges (and the upcoming
imported_by resolver) can now import ModuleSentinel without pulling in
the full inspect module.

All 79 existing inspect tests pass unmodified (git diff of test file is
empty). No circular imports confirmed. Full suite: 1784 passed, 86.6%
coverage (above 85% threshold). The 2 failures are the known
performance-flaky tests under coverage load.
